Made diagnostic messages from is and isnt mention whether this is a todo test
The Perl diagnostics have language like "Failed (TODO) test" when
failing a todo test.  I added this to libtap++.

To add it, I had to make the TODO status visible to the is and isnt
methods, which are defined in the header file.  I thought that
TAP::details was the best place to put the interface function -- out
of the way of most users, but still where is and isnt can see it.